% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_abundance}
\alias{compute_abundance}
\title{Compute the abundance scores of nodes in a network}
\usage{
compute_abundance(dfn, node, ids, expr, group, cois, gois, byImmune = FALSE)
}
\arguments{
\item{dfn}{A dataframe with the cell abundance and gene expression values, in a tidy format.
Each row of the measurement data should contain, at least, the following columns:
- a sample ID
- the node
- the expression value
- group of this sample ID}

\item{node}{Name of the column with the nodes names.}

\item{ids}{Name of the column with the ID of each sample.}

\item{expr}{Name of the column with the expression data.}

\item{group}{Name of the column with the group information of the sample.}

\item{cois}{Cells of interest to filter from the scaffold.}

\item{gois}{Genes of interest to be filtered.}
}
\value{
A dataframe with the ternary bins for each node and the ratio of sample in each group that map to the two upper tertiles.
}
\description{
Compute the abundance scores of nodes in a network
}
