% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_concordance}
\alias{compute_concordance}
\title{Compute the concordance scores of nodes in a network}
\usage{
compute_concordance(scaffold, nodes_info)
}
\arguments{
\item{scaffold}{A dataframe with a scaffold interaction in each row. Columns should be named "From" and "To".}

\item{nodes_info}{A dataframe with the ternary bins for each node and the ratio of samples in each group that map to the two upper tertiles. Use the output dataframe from compute_abundance function.}
}
\value{
A dataframe with the concordance score for each edge in the scaffold.
}
\description{
Compute the concordance scores of nodes in a network
}
