% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{filter_network}
\alias{filter_network}
\title{Filter the edges table to nodes above a given abundance ratio and edges above a given concordance level.}
\usage{
filter_network(nodes_scores, edges_scores, abundance, concordance)
}
\arguments{
\item{nodes_scores}{A dataframe with abundance score for all nodes across all groups. Use the output dataframe from get_abundance_table function.}

\item{edges_scores}{A dataframe with the concordance score for each edge in the scaffold. Use the output dataframe from compute_concordance function.}

\item{abundance}{The abundance threshold to filter nodes. Should be between 0 and 1.}

\item{concordance}{The concordance threshold to filter edges. Only positive values.}
}
\value{
A dataframe with all nodes with UpBinRatio above the abundance threshold and edges with ratioScores above the concordance threshold.
}
\description{
Filter the edges table to nodes above a given abundance ratio and edges above a given concordance level.
}
