% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{get_abundance_table}
\alias{get_abundance_table}
\title{Consolidate the nodes abundance score in a dataframe ready to be saved to file.}
\usage{
get_abundance_table(nodes_info)
}
\arguments{
\item{nodes_info}{A dataframe with the ternary bins for each node and the ratio of samples in each group that map to the two upper tertiles. Use the output dataframe from compute_abundance function.}
}
\value{
A dataframe with abundance score for all nodes across all groups.
}
\description{
Consolidate the nodes abundance score in a dataframe ready to be saved to file.
}
