% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{get_scaffold}
\alias{get_scaffold}
\title{Filter edges from a scaffold based on cells and genes of interest.}
\usage{
get_scaffold(scaffold, cois, gois)
}
\arguments{
\item{scaffold}{A dataframe with at least two columns. The two first columns should describe the edges - ie. the first column being the source node and the second the target node.}

\item{cois}{Cells of interest to filter from the scaffold.}

\item{gois}{Genes of interest to be filtered.}
}
\value{
A dataframe cointaing the edges in the scaffold that have the cois connected to gois, and edges containig at least one gois.
}
\description{
Filter edges from a scaffold based on cells and genes of interest.
}
